test-script/pyproject.toml
[project]
name = "locust-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "locust>=2.42.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
]
